feat : responsiveness
