+++ src/components/Sidebar/Sidebar.css
@@ -2,10 +2,11 @@
   margin-top: 1rem;
   width: 14rem;
   position: fixed;
-  border-right: 1px solid var(--secondary-color);
-  padding: 0.4rem 0;
+  padding: 5px 0;
   overflow-y: scroll;
-  max-height: 25rem;
+  max-height: 30rem;
+  z-index: 10;
+  background-color: var(--secondary-color);
 }
 .filter-btn,
 .filter-text {
@@ -59,4 +60,15 @@
   margin: 1rem;
   font-style: italic;
   font-weight: bold;
+}
+
+@media screen and (max-width: 792px) {
+ .sidebar{
+  margin-top: 0;
+  width: 100%;
+  left: 0;
+  right: 0;
+  height: 20rem;
+  border: 2px solid var(--primary-color);
+ }
 }
