+++ src/index.css
@@ -1,0 +1,9 @@
+html{
+    font-size: 100%;
+}
+
+@media screen and (max-width: 792px) {
+ html{
+    font-size: 70%;
+ }
+}
