+++ src/pages/CartPage/CartPage.css
@@ -7,3 +7,9 @@
     position: fixed;
 }
 
+/* @media screen and (max-width: 792px) {
+ .cart-page{
+    flex-direction: column;
+ }
+} */
+
